demos fixed with os paths to library #35
