+++ Wrappers/Python/Demos/DemoModel3D.py
@@ -12,18 +12,22 @@ If one needs to modify/add phantoms, please edit Phantom3DLibrary.dat
 @author: Daniil Kazantsev
 """
 import timeit
+import os
 from tomophantom import TomoP3D
 import matplotlib.pyplot as plt
+import tomophantom
 
 print ("Building 3D phantom using TomoPhantom software")
 tic=timeit.default_timer()
-model = 2
+model = 2 # select a model number from the library
 # Define phantom dimensions using a scalar (cubic) or a tuple [N1, N2, N3]
 N_size = 256 # or as a tuple of a custom size (256,256,256)
-#specify a full path to the parameters file
-pathTP3 = '../../../PhantomLibrary/models/Phantom3DLibrary.dat'
+# one can specify an exact path to the parameters file
+# path_library2D = '../../../PhantomLibrary/models/Phantom3DLibrary.dat'
+path = os.path.dirname(tomophantom.__file__)
+path_library3D = os.path.join(path, "Phantom3DLibrary.dat")
 #This will generate a N_size x N_size x N_size phantom (3D) or non-cubic phantom
-phantom_tm = TomoP3D.Model(model, N_size, pathTP3)
+phantom_tm = TomoP3D.Model(model, N_size, path_library3D)
 toc=timeit.default_timer()
 Run_time = toc - tic
 print("Phantom has been built in {} seconds".format(Run_time))
@@ -48,6 +52,8 @@ plt.show()
 import timeit
 from tomophantom import TomoP3D
 import matplotlib.pyplot as plt
+import tomophantom
+import os
 
 print ("Building a subset of 3D phantom using TomoPhantom software")
 tic=timeit.default_timer()
@@ -55,10 +61,10 @@ model = 3
 # Define phantom dimensions using a scalar (cubic) or a tuple [Z, Y, X]
 DIM = (256,256,256) # full dimension of required phantom (z, y, x)
 DIM_z = (94, 158) # selected vertical subset (a slab) of the phantom
-#specify a full path to the parameters file
-pathTP3 = '../../functions/models/Phantom3DLibrary.dat'
+path = os.path.dirname(tomophantom.__file__)
+path_library3D = os.path.join(path, "Phantom3DLibrary.dat")
 #This will generate a N1 x N2 x N_slab phantom (3D)
-phantom_tm = TomoP3D.ModelSub(model, DIM, DIM_z, pathTP3)
+phantom_tm = TomoP3D.ModelSub(model, DIM, DIM_z, path_library3D)
 #phantom_tm = TomoP3D.Model(model, DIM, pathTP3)
 toc=timeit.default_timer()
 Run_time = toc - tic
