+++ Wrappers/Python/Demos/DemoModel3D_t.py
@@ -4,7 +4,7 @@
 GPLv3 license (ASTRA toolbox)
 Note that the TomoPhantom package is released under Apache License, Version 2.0
 
-Script to generate 4D analytical phantoms (wip: generation of 4D projection data )
+Script to generate 4D (3D+time) analytical phantoms (wip: generation of 4D projection data )
 If one needs to modify/add phantoms, please edit Phantom3DLibrary.dat
 
 !Run script from "Demos" folder in order to ensure a correct path to *dat file!
@@ -12,19 +12,23 @@ If one needs to modify/add phantoms, please edit Phantom3DLibrary.dat
 @author: Daniil Kazantsev
 """
 import timeit
-from tomophantom import TomoP3D
+import os
 import matplotlib.pyplot as plt
 import numpy as np
+import tomophantom
+from tomophantom import TomoP3D
 
 print ("Building 4D phantom using TomoPhantom software")
 tic=timeit.default_timer()
-model = 100
+model = 100 # note that the selected model is temporal (3D + time)
 # Define phantom dimensions using a scalar (cubic) or a tuple [N1, N2, N3]
 N_size = 256 # or as a tuple of a custom size (256,256,256)
-#specify a full path to the parameters file
-pathTP3 = '../../../PhantomLibrary/models/Phantom3DLibrary.dat'
+# one can specify an exact path to the parameters file
+# path_library2D = '../../../PhantomLibrary/models/Phantom3DLibrary.dat'
+path = os.path.dirname(tomophantom.__file__)
+path_library3D = os.path.join(path, "Phantom3DLibrary.dat")
 #This will generate a Time x N_size x N_size x N_size phantom (4D)
-phantom_tm = TomoP3D.ModelTemporal(model, N_size, pathTP3)
+phantom_tm = TomoP3D.ModelTemporal(model, N_size, path_library3D)
 toc=timeit.default_timer()
 Run_time = toc - tic
 print("Phantom has been built in {} seconds".format(Run_time))
@@ -48,8 +52,10 @@ for i in range(0,np.size(phantom_tm,0)):
     plt.pause(0.3)
 
 #%%
-# The capability of building a subset of vertical slices out of 4D phantom (faster)
+# A capability of building a subset of vertical slices out of 4D phantom (faster)
 import timeit
+import os
+import tomophantom
 from tomophantom import TomoP3D
 import matplotlib.pyplot as plt
 
@@ -59,10 +65,10 @@ model = 101
 # Define phantom dimensions using a scalar (cubic) or a tuple [Z, Y, X]
 DIM = (256,256,256) # full dimension of required phantom (z, y, x)
 DIM_z = (94, 158) # selected vertical subset (a slab) of the phantom
-#specify a full path to the parameters file
-pathTP3 = '../../functions/models/Phantom3DLibrary.dat'
+path = os.path.dirname(tomophantom.__file__)
+path_library3D = os.path.join(path, "Phantom3DLibrary.dat")
 #This will generate a Time x N1 x N2 x N_slab phantom (4D)
-phantom_tm = TomoP3D.ModelTemporalSub(model, DIM, DIM_z, pathTP3)
+phantom_tm = TomoP3D.ModelTemporalSub(model, DIM, DIM_z, path_library3D)
 #phantom_tm = TomoP3D.Model(model, DIM, pathTP3)
 toc=timeit.default_timer()
 Run_time = toc - tic
